Feature: Add Query features
In porting this feature from the marketplace,
I decided to keep the graphql client that is
used in that codebase. This client generates the
graphql queries and mutations from the schema
automatically and because of that, I believe that
we can replace that library with the preexisting
one for faster development in the future. I was unable
to do that refactor in this PR. We can use issue #2 to
track that work.